JBRULES-1958 Improve JPA persistence
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@25445 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-core/src/main/java/org/drools/common/InternalWorkingMemory.java
@@ -28,9 +28,9 @@ import org.drools.time.TimerService;
 public interface InternalWorkingMemory
     extends
     WorkingMemory {
-    public long getId();
+    public int getId();
     
-    public void setId(long id);
+    public void setId(int id);
     
     void setRuleBase(final InternalRuleBase ruleBase);
